requirements.txt
accelerate==0.26.1
alembic==1.13.1
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
Authlib==1.3.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.30.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
boilerpy3==1.0.7
boto3==1.34.34
botocore==1.34.34
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.28.1
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
databricks-cli==0.18.0
distro==1.9.0
dnspython==2.5.0
docker==7.0.0
docopt==0.6.2
elastic-transport==7.16.0
elasticsearch==7.17.9
entrypoints==0.4
Events==0.5
exceptiongroup==1.2.0
faiss-cpu
farm-haystack==1.24.0
fastapi==0.108.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
haystack-bm25==1.0.2
haystack_threshold_node==0.0.1
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
inflect==7.0.0
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langdetect==1.0.9
lazy-imports==0.3.1
lit==17.0.6
llvmlite==0.42.0
loguru==0.7.2
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
mlflow==2.10.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msrest==0.7.1
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numba==0.59.0
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.9.4.25
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.11.0
openai-whisper==20231117
opensearch-py==2.4.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pillow==10.2.0
pinecone-client==2.2.4
platformdirs==4.2.0
portalocker==2.8.2
posthog==3.3.4
prompthub-py==4.0.0
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==15.0.0
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
pymongo==4.6.1
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pynvml==11.5.0
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-frontmatter==1.1.0
python-magic==0.4.27
python-multipart==0.0.7
python-pptx==0.6.23
pytz==2023.4
PyYAML==6.0.1
qdrant-client==1.7.2
qdrant-haystack==1.0.11
quantulum3==0.9.0
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
rest-api @ git+https://github.com/deepset-ai/haystack.git@d2ca9c94e02f7245ae3b9c6738e2695658eaa2e9#subdirectory=rest_api
rpds-py==0.17.1
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
selenium==4.17.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.51
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
sseclient-py==1.8.0
starlette==0.32.0.post1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tika==2.6.0
tiktoken==0.5.2
tokenizers==0.15.1
torch==2.0.0
tqdm==4.66.1
transformers==4.36.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==1.26.18
uvicorn==0.27.0.post1
validators==0.22.0
weaviate-client==3.26.2
Werkzeug==3.0.1
wsproto==1.2.0
XlsxWriter==3.1.9
zipp==3.17.0
